requirements.txt
annotated-types==0.7.0
certifi==2024.12.14
charset-normalizer==3.4.1
click==8.1.8
cloudpickle==3.1.0
docstring_parser==0.16
grpcio==1.68.1
idna==3.10
jsonref==1.1.0
markdown-it-py==3.0.0
mdurl==0.1.2
networkx==3.4.2
numpy==2.2.1
opencv-python-headless==4.10.0.84
pathlib==1.0.1
protobuf==4.25.5
pydantic==2.10.4
pydantic_core==2.27.2
Pygments==2.18.0
python-dateutil==2.9.0.post0
PyYAML==6.0.2
requests==2.32.3
rich==13.9.4
shellingham==1.5.4
sievedata==1.4.7
six==1.17.0
sseclient==0.0.27
tqdm==4.67.1
typeguard==4.4.1
typer==0.15.1
typing_extensions==4.12.2
urllib3==2.3.0
